refactor: move analyze capability to the struct
